API-8140: rename file and move controllers as per PR-bot
